Improve denoising and color thresholding
Now takes in a CSV as colors to map image colors to. Improved the
connected-component labelling implementation to correctly join regions
and also doesn't just merge all colors based on the most popular one in
the region. This may prove problematic for the Parrot but that can be a
toggle. Added a median filter to denoise. Works best on random spot
noise but has the added benefit of effectively preserving edges.